requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
astroid==2.4.2
backcall==0.2.0
certifi==2021.10.8
colorama==0.4.4
debugpy==1.3.0
decorator==5.0.9
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
Markdown==3.3.7
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.19
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pylint==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==301
pyzmq==22.1.0
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.0
